Version/v3.0

+++ src/commons/data-managers/tenant-manager.js
@@ -1,56 +1,18 @@
-const validate = require("jsonschema").validate;
-
 const Tenant = require("../entities/tenant");
-const dbm = require("../utilities/database-manager");
-const SecurityUtils = require("../utilities/security-utils");
-
-const TENANT_ENCRYPT_KEYS = [
-  "paymentMerchantId",
-  "paymentProjectId",
-  "paymentSecret",
-  "noreplyPassword",
-  "password",
-  "noreplyGraphClientSecret",
-];
+const TenantModel = require("./models/tenantModel");
 
 /**
  * Data Manager for Tenant objects.
  */
 class TenantManager {
   /**
-   * Check if an object is a valid Tenant.
-   *
-   * @param {object} tenant A tenant object
-   * @returns true, if the object is a valid tenant object
-   */
-  static validateTenant(tenant) {
-    const schema = require("../schemas/tenant.schema.json");
-    return validate(tenant, schema).errors.length === 0;
-  }
-
-  /**
    * Get all tenants
    * @returns List of tenants
    */
-  static getTenants() {
-    return new Promise((resolve, reject) => {
-      dbm
-        .get()
-        .collection("tenants")
-        .find({})
-        .toArray()
-        .then((rawTenants) => {
-          const tenants = rawTenants.map((rt) => {
-            const tenant = new Tenant(rt);
-            tenant.applications = tenant.applications.map((app) => {
-              return SecurityUtils.decryptObject(app, TENANT_ENCRYPT_KEYS);
-            });
-            return SecurityUtils.decryptObject(tenant, TENANT_ENCRYPT_KEYS);
-          });
-
-          resolve(tenants);
-        })
-        .catch((err) => reject(err));
+  static async getTenants() {
+    const rawTenants = await TenantModel.find();
+    return rawTenants.map((rt) => {
+      return new Tenant(rt);
     });
   }
 
@@ -60,24 +22,12 @@ class TenantManager {
    * @param {string} id Logical identifier of the bookable object
    * @returns A single bookable object
    */
-  static getTenant(id) {
-    return new Promise((resolve, reject) => {
-      dbm
-        .get()
-        .collection("tenants")
-        .findOne({ id: id })
-        .then((rt) => {
-          if (!rt) {
-            return reject(new Error(`No tenant found with ID: ${id}`));
-          }
-          const tenant = new Tenant(rt);
-          tenant.applications = tenant.applications.map((app) => {
-            return SecurityUtils.decryptObject(app, TENANT_ENCRYPT_KEYS);
-          });
-          resolve(SecurityUtils.decryptObject(tenant, TENANT_ENCRYPT_KEYS));
-        })
-        .catch((err) => reject(err));
-    });
+  static async getTenant(id) {
+    const rawTenant = await TenantModel.findOne({ id: id });
+    if (!rawTenant) {
+      return null;
+    }
+    return new Tenant(rawTenant);
   }
 
   /**
@@ -88,20 +38,14 @@ class TenantManager {
    * @returns Promise<>
    */
   static async storeTenant(tenant, upsert = true) {
-    try {
-      const tenantsCollection = dbm.get().collection("tenants");
-      tenant.applications = tenant.applications.map((app) => {
-        return SecurityUtils.encryptObject(app, TENANT_ENCRYPT_KEYS);
-      });
+    const newTenant = new Tenant(tenant);
 
-      await tenantsCollection.replaceOne(
-        { id: tenant.id },
-        SecurityUtils.encryptObject(tenant, TENANT_ENCRYPT_KEYS),
-        { upsert: upsert },
-      );
-    } catch (err) {
-      throw new Error(`Error storing tenant: ${err.message}`);
-    }
+    await TenantModel.updateOne({ id: tenant.id }, newTenant, {
+      upsert: upsert,
+      setDefaultsOnInsert: true,
+    });
+
+    return newTenant;
   }
 
   /**
@@ -110,63 +54,54 @@ class TenantManager {
    * @param {string} id The identifier of the tenant
    * @returns Promise<>
    */
-  static removeTenant(id) {
-    return new Promise((resolve, reject) => {
-      dbm
-        .get()
-        .collection("tenants")
-        .deleteOne({ id: id })
-        .then(() => resolve())
-        .catch((err) => reject(err));
-    });
+  static async removeTenant(id) {
+    await TenantModel.deleteOne({ id: id });
   }
 
   static async getTenantApps(tenantId) {
-    try {
-      const tenant = await dbm.get().collection("tenants").findOne({
-        id: tenantId,
-      });
-      const applications = tenant.applications;
-      return applications.map((app) => {
-        return SecurityUtils.decryptObject(app, TENANT_ENCRYPT_KEYS);
-      });
-    } catch {
-      throw new Error(`No tenant found with ID: ${tenantId}`);
-    }
+    const rawTenant = await TenantModel.findOne({ id: tenantId });
+    const tenant = new Tenant(rawTenant);
+    return tenant.applications;
   }
 
   static async getTenantApp(tenantId, appId) {
-    try {
-      const tenant = await dbm.get().collection("tenants").findOne({
-        id: tenantId,
-      });
-      const application = tenant.applications.find((app) => app.id === appId);
-      return SecurityUtils.decryptObject(application, TENANT_ENCRYPT_KEYS);
-    } catch {
-      throw new Error(`No tenant found with ID: ${tenantId}`);
-    }
+    const rawTenant = await TenantModel.findOne({ id: tenantId });
+
+    const tenant = new Tenant(rawTenant);
+    return tenant.applications.find((app) => app.id === appId);
   }
 
   static async getTenantAppByType(tenantId, appType) {
-    try {
-      const tenant = await dbm.get().collection("tenants").findOne({
-        id: tenantId,
-      });
+    const rawTenant = await TenantModel.findOne({ id: tenantId });
 
-      const applications = tenant.applications.filter(
-        (app) => app.type === appType,
-      );
-      return applications.map((app) => {
-        return SecurityUtils.decryptObject(app, TENANT_ENCRYPT_KEYS);
-      });
-    } catch {
-      throw new Error(`No tenant found with ID: ${tenantId}`);
-    }
+    const tenant = new Tenant(rawTenant);
+    return tenant.applications.filter((app) => app.type === appType);
   }
+
   static async checkTenantCount() {
     const maxTenants = parseInt(process.env.MAX_TENANTS, 10);
-    const count = await dbm.get().collection("tenants").countDocuments({});
+    const count = await TenantModel.countDocuments({});
     return !(maxTenants && count >= maxTenants);
+  }
+
+  static async getTenantUsers(tenantId) {
+    const rawTenant = await TenantModel.findOne({ id: tenantId });
+    if (!rawTenant) {
+      return null;
+    }
+    const tenant = new Tenant(rawTenant);
+    return tenant.users;
+  }
+
+  static async getTenantUsersByRoles(tenantId, roles) {
+    const rawTenant = await TenantModel.findOne({ id: tenantId });
+    if (!rawTenant) {
+      return [];
+    }
+    const tenant = new Tenant(rawTenant);
+    return tenant.users.filter((user) =>
+      user.roles.some((role) => roles.includes(role)),
+    );
   }
 }
